renamed small perf method as sanity (which they are) removed stress module, since none of methods in ther really do a stresss test. Introduced sanity module which now take its responsibility. Added soak test for 48 hours
