+++ jormungandr-integration-tests/Cargo.toml
@@ -61,8 +61,9 @@ lazy_static = "1.3"
 ed25519-bip32 = "0.1"
 
 [features]
-soak-test = []
 testnet = []
+sanity-non-functional = []
+soak-non-functional = []
 
 [build-dependencies]
 protoc-rust = "2.0"
